requirements.txt
# Core dependencies - AI/ML
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Hebrew voice recognition
librosa>=0.10.0
sounddevice>=0.4.6
soundfile>=0.12.0
SpeechRecognition>=3.10.0
openai-whisper>=20230314

# Hebrew TTS
pygame>=2.5.0
gtts>=2.3.2
edge-tts>=6.1.0
requests>=2.31.0

# AI and LLM
groq>=0.4.0
openai>=1.0.0

# Google Sheets integration
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Computer vision (for food analysis)
opencv-python>=4.8.0
Pillow>=10.0.0

# Automation and utilities
pyautogui>=0.9.54
keyboard>=0.13.5
schedule>=1.2.0

# Optional: Instagram automation
# instapy>=0.6.16
# selenium>=4.11.0

# Development tools
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0